+++ R/ggaverage.R
@@ -1,6 +1,5 @@
 #' @rdname ggpredict
 #' @importFrom dplyr select
-#' @importFrom tibble as_tibble
 #' @importFrom rlang .data
 #' @export
 ggaverage <- function(model, terms, ci.lvl = .95, type = c("fe", "re"), typical = "mean", ppd = FALSE, x.as.factor = FALSE, pretty = TRUE, condition = NULL, ...) {
@@ -27,10 +26,6 @@ ggaverage <- function(model, terms, ci.lvl = .95, type = c("fe", "re"), typical 
   }
 
 
-  # to tibble
-  zus <- tibble::as_tibble(zus)
-
-
   # add back attributes. therefore, we first copy the attributes from the
   # original data frame and delete the common attributes, like class etc.
   # and then add attributes to our final df
@@ -53,10 +48,9 @@ ggaverage <- function(model, terms, ci.lvl = .95, type = c("fe", "re"), typical 
 # As categorical variables do not necessarily need to follow a "linear"
 # (or other distribution based) trend, we can simply calculate the mean
 #' @importFrom dplyr group_by summarise ungroup
-#' @importFrom tibble has_name
 get_average_values <- function(dat) {
   # do summary, depending on third group
-  if (tibble::has_name(dat, "facet")) {
+  if (obj_has_name(dat, "facet")) {
     dat <- dplyr::group_by(dat, .data$x, .data$group, .data$facet)
   } else {
     dat <- dplyr::group_by(dat, .data$x, .data$group)
@@ -72,7 +66,7 @@ get_average_values <- function(dat) {
     dplyr::ungroup()
 
   # sort columns
-  if (tibble::has_name(dat, "facet")) {
+  if (obj_has_name(dat, "facet")) {
     zus <- zus[, c(1, 4:6, 2:3)]
   } else {
     zus <- zus[, c(1, 3:5, 2)]
@@ -85,14 +79,13 @@ get_average_values <- function(dat) {
 # this method prepares the data to get smoothed predictions for
 # average effects
 #' @importFrom dplyr group_by summarise ungroup arrange mutate select
-#' @importFrom tibble has_name
 #' @importFrom tidyr nest unnest
 #' @importFrom purrr map
 #' @importFrom stats predict loess lm
 #' @importFrom sjmisc var_rename
 get_smoothed_avg <- function(dat) {
   # do summary, depending on third group
-  if (tibble::has_name(dat, "facet")) {
+  if (obj_has_name(dat, "facet")) {
     tmp <- dplyr::group_by(dat, .data$x, .data$group, .data$facet)
   } else {
     tmp <- dplyr::group_by(dat, .data$x, .data$group)
@@ -103,7 +96,7 @@ get_smoothed_avg <- function(dat) {
     dplyr::summarise(predicted = mean(.data$predicted)) %>%
     dplyr::ungroup()
 
-  if (tibble::has_name(dat, "facet")) {
+  if (obj_has_name(dat, "facet")) {
     zus <- zus %>%
       dplyr::arrange(.data$x, .data$group, .data$facet) %>%
       dplyr::group_by(.data$group, .data$facet)
@@ -162,7 +155,7 @@ get_smoothed_avg <- function(dat) {
     sjmisc::var_rename(avgpred = "predicted")
 
   # re-arrange columns
-  if (tibble::has_name(dat, "facet"))
+  if (obj_has_name(dat, "facet"))
     zus <- zus[, c(4, 3, 5:6, 1:2)]
   else
     zus <- zus[, c(3, 2, 4:5, 1)]